requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.28
aiosqlite>=0.20.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
apscheduler>=3.10.4
google-generativeai>=0.4.1
yfinance>=0.2.37
httpx>=0.27.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
multipart>=0.0.5
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
